refactor: move callbacks to separate module
